Move code from helper location to the class that actually uses it.
